Springs now implemented
Remember that super secret commented line in physics.c. Well turns out I forgot to zero out the link force and caused the whole thing to resonate out of existance. Now it tends to keep resonating. Which is cool. Actually it's hot when you think about it, but besides that terrible pun, I need to work on the link system. Probably going to add a standard array to store all the objects linked to an object along with an int which stores how many links the object's having. Should make everything so much easier down the road.