Memory hole most definately fixed
Okay, so the memory hole has been fixed. For a second time. I knew it was SDL. It had to be. Turns out it's SDL_FreeSurface(imgTxt);. Why on earth does it not free everything? I think it might be possible to reduce this code down, but man, dealing with the memory loss (read: rm -rf ~ or who the hell makes a dir called ~ in their ~ directory) and trying to think of a way to make bonding more rigid is kinda hard. I think it's better if I go for a doublesided bonds, which means every atom in a bond will interact with always interact with other atoms and vice versa. Also will explore trying to realistically model molecular interactions (read: molecular collisions). Should be just a matter of adjusting constants and multiplying dimensions(when rendering so we can see what's going on).