Updated object initialization
Just switched to using v4sf, saves us a few lines. Still thinking on what I should do next. 3D OpenGL's apparently quite hard to do, and with the broken mouse support for SDL2, not sure how to do it. So perhaps something to do with the physics engine. A parser and a better kind of normalization sound good right now.