Switched to long double
Still not sure if it was the right thing to do. Need to do performance measurements. But I mean we should still get to use the FPU of the main processor and not do them all in software, besides this is only for the constants, not for the positions, so shouldn't be so bad