Parser added
Yup, working on a parser right now. Not working, but compiles fine-ish. Using a really roundabout way of doing it as well. Which I just realized does not work. Can't handle double digits for number of links. Maybe if I split it into rows it'd be fine, though I'd much rather keep everything about object i on a single row.