Big code cleanup
TTF removed because it was crap. Didn't free the damn texture. It has to be a bug with SDL TTF. Screw it. ncurses is the future. Or just freetype. Or maybe... libass. That aside, I must stay away from whatever I was on. How else does one scre up the pow() function that badly? Now that the program's light as it can be, I think it's time we gave this 3D gimmick a try. For real.