Temporarily disabled FPS display
Not given up yet. Did some tracing and it turns out X redraws the window every time it's title is updated. Notthing wrong with that, not at all. Just not okay if you tend to do this thousands of times per second. Not sure if the issue persists if compositing is enabled, but anyway, to the glorious land of Freetype I shall sail. Heard it's nice and hot this time of the year there. There's apparently plenty of booty to gain ownership of as well. Anyway, day 2 of filler commits is done. Secret project not going well at all. Will work on getting that one fixed by new year. Also, if anyone reading this is wondering what to listen to, google the string 'machine jam psi'. Thank me later.

>Implying anyone is reading this