Updated comment in posdata.dat
Hang in there, almost a week left. No work done on super secret project today. Still needs tweaking. Why does it freeze when periodic boundaries are enabled seems to be a problem. Not to worry, should be done by Monday. And then a week left to do additional tweaks and a report. All cool. Still unsure where to go with physengine. What do I do? I want to get 3D up and running. The physics.c file is already almost fully N-dimensional. Not going to bother with FPS counting. Maybe I'll add a parameter to disable any visual stuff so it can do it's job as fast as possible. Enough rambling for today.