main: implemented a (working) FPS counter, handling QUIT in a new way
Not the most efficient FPS counting method one could think of, but it works. Have an idea on how to make it better though. And yes, our code has been infected with a goto. No, no, don't point that gun at me, I'm just using it as a way to handle quitting.