main: removed non-functioning SDL event code
HEY, A MAJOR MILESTONE! I BROKE THE X SERVER. What do you mean it's not good to pick on buggy code?

Anyway, removed SDL code completely. It was 'it's my video window or the highway' for it, and it didn't support input through an alien window. I'm researching something which will allow me to grab input better. Perhaps I'll implement it in a seperate function for each video output and call whichever it's been compiled/called with. All code's backup'd so I'll do some copypasting as I imagine there isn't going to be much difference.