glfun: added additional rotation matrices
Still, what you're seeing is not a projection. It's still 2D, except we can rotate the 2D image, if that makes any sense. Have to look into exactly what's up. Also, that new folder in resources? Yup, it's coming up soonishly. Probably lots of work to do on the parser as well.