main, glfun: GL_DEPTH_TEST enabled, fixed memory leak, reset scale
Was I on something when I wrote that? Did I seriously think such a small memory hole was pathetic compared to the scale of the Titanic? What if someone was to run this for hours just resizing? Blasphemy. Anyway, totally forgot to enable depth. Glad I didn't have to tinker.