pthreads: rewrote and finalized threading, CPU measurement, --std=g99
'Hey just how many times are you going to rewrite that damn thing?'

Until it's fine enough that I'd consider it useable. Like right now. It's not perfect, but it's fine. Especially if you've got a uniform distribution of objects around the system. Each thread has a set of objects to calculate. And each thread has to have finished moving all of its objects, or otherwise you'd be getting inaccuracies. Here I used a mutex to put a lock until every thread has reached that point. I also use it as a way to pause the engine without shutting the threads down. Which is awesome. Threads now no longer need to be initialized every single time or object, now we can just keep 'em running until the cows come home. And as a bonus, using --novid will now result in the main thread just waking up only 10 times a second(SDL_Delay(100)). Unfortunately had to switch standards to C99 + GNU extensions. Which isn't really unfortunate because both Clang and GCC(duh) support (and both use it by default) this standard. C11 (not this C++11 junk) seems to have interesting stuff in it, so I'd probably switch once Travis-cl updates compilers (not likely to happen). Anyway, that's it for threading I believe. I need to get the LJ potential up and running, which I am working on, get a better display and a file reader for molecules and tweak the linking so that it's dynamic by some predetermined condition. I think it'd be really hard to get a shaded spheres though.