main, physics: define fps var as volatile, do not spawn a thread if ptr!=NULL
Still need to seperate the thread object distribution to a seperate function. But for now, the FPS issue is fixed. Turns out GCC likes to optimize it for some reason. Assembly depricated? I'd say it's pretty damn useful when you need to see what the compiler's doing. Especially with GCC.

Also, in case anyone from the future's writing a report on 21st century's Desktop Environments, go with KDE. Took me less than 2 hours to fully set up compared to the few days for GNOME. It's so much faster as well. Who'd have thought JavaScript does not do well in time-critical applications. Seriously, running an entire JavaScript interpreter for a calendar/network manager/menu. And those options... KDE has so many options it's impossible not to have it set up exactly the way you want it to be. Though I'm still a fan of GTK+3.