glfun: spherical rendering enabled
Took me a while. Turns out not many people want to render spheres, naturally. Heck, I even thought of doing a hack and exploiting the U(1) group by using a sprite. But I just took a look at what should be a very familiar integral for the reader and decided to do it myself. It's not exactly a sphere, but looks close enough at a distance. Will improve later.