molreader, parser: implemented basic PDB file importing.
It's a start. PDB files contain a lot of text and reading is unreliable. I have only tested them with the 1gcn molecule/protein. I absolutely need to find a way to read files and ignore the spaces between the numbers/strings. Need to research exactly how to implement residues and temperature factor and occupancy(?) and all the other stuff, but it works for now.