physics, README: updated min compiler versions, error handling
Well, I did go and test my program on a linux mint. It didn't work with their repos. Compiling with GCC caused segfaults in the assignment code and overflows in reading the periodic table. Clang completely refused to acknowledge the assignment of values to object's pos arrays. It was as if that was an alien concept. Taugtology. I gave up on using silly abstractions realizing I was about to miss my bus and said to hell with it. Update your damn compilers or expect broken stuff. Still using Ubuntu 12.04? Unacceptable. My laptop's newer than that. And it has a Celeron, so that's something.