all: use C11+GNU std, Clang workaround, color coding messages
Tests today went fine. All seems almost ready for release and next week is the week to do so. Clang's having problems with V4SD assignmnet but hey, I only save 2 lines, might as well do it verbosely. Also, input file is a better name. Because that's all posdata is. Anyone who's worked with LAMMPS or TOWHEE or whatever knows it so might as well stick to the standard.