Version 0.1 official release
After months of hard work and commits every single day(for a period of 2 months) here it is. It's far from perfect, but what is? I've got exams to prepare for, so for a month and a half or so things will stay quiet. If I ever get pull requests, I'll be sure to pay attention. I'll also be hanging around the IRC all the time(irssi on a server, whoa) and I always check my email. Hopefully will at least get some hate mail from python occultists or C++ fanatics. There's that and that's that. Over and out.