main: use getopt() for flag detection
Works as well as my previous code. That means it's as stable as my last code's. That means not very stable. I have not figured out exactly how to set the conditions. Also, I plan on moving to system integration as soon as I get fontconfig working. So you could just type "physengine -f file" in a terminal and do whatever you need to do. I'll need to merge the simconf.conf file and the posdata.in file as well. Will work on that as well as getting a quadtree algorithm up and running.