lua: ignore NULL objects, cleanup root
Took me a week and I still couldn't figure out where that bloody object 0 came from or why it even counted. Anyway, we now ignore it. It's a very very hacky solution and I don't like it one bit, but it works. Now you can reenable forces and we're back to feature parity with old system.