physics: make the thread settings live on the stack
It's pointless to keep tossing around pointers because of how long it might take to dereference them. And we do that quite often inside the thread. So instead now just spawn a variable and load it with the structure we sent. I'll probably move the thread_opts struct inside and handle thread time statistics entirely inside the main function. I can start to get rolling real fast as my exams are over. I've got big plans as well.