physics-bh: fixed GCC scalar->vector assignment in octree generation
"I-It's Clang's fault! It did the assignment like it was no one's business"

But seriously, I like the way GCC deals with this. Sure it may be stupid not to do the assignment implicitly but it's foolproof. Clang's quite lax with the warnings, even when using --pedantic and -Wall(which we always are here, and you should too).