main, physics: use FontConfig to choose fonts, BH tree cleanup f-n
You don't need a good education in Group Theory to use fontconfig but it helps. You'll probably need a PhD in Group Theory if you want to mess around with it though. But that's some damn good library, I'll tell you what. I dislike the "black-box" FcPattern but oh well, apparently it's a structure too complex for us human beings to comprehend.

I'm still on the fence how to do Octree generation and creation. My current plan is to log all BH trees for a time period proportional with the highest speed of an object and the time constant and just check how often that cell is filled and in case it has not been used for the entire lenght, free it. Seems like a reasonable tradeoff between calls to malloc and performance. Because you'd really want to avoid calling malloc trillions of times every look repetition. Sure, it's all in virtual space anyhow, but I'd rather malloc 8 cells at a time rather than 8 seperate mallocs, you know?

Version 0.2 is coming up real soon. As soon as I have a good enough BH algorithm implemented and I have sorted out the mess. In case the shaders prove too difficult/messy to put in the filesystem I'll just embed them inside the headers somewhere. I'll have to clean the git history from all the molecule and font files too. Sometime down the road maybe. Before 0.2. As to what'll be in 0.3? We'll hopefully go full quantum.