all: renamed elements to phys_misc, numbers f-n moved, housekeeping
Things are looking grim. No, the project's going just fine, need to do a bit of tweaks to the octree generation and make the thread function work and we'd be set for 0.2. But the time I'll be able to devote will be limited until I find a job and settle down for the summer. I've got no options(had a possibility until today, then that possibility went south) and I still have 2 more options from which I've yet to get an email. So for a few more weeks we'll be on the backburner, possibly more depending on what kind of  challenge calls out to me.