graph: modularized graphics into multiple files
"And so it came to be that the once great empire of GLFUN was cast asunder into areas with different flags..."

This will make it easy as hell to be able to compile the engine without any graphics/fonts and make it standalone. Still a couple of bugs and missing features(to plumb or to move?). I now need to man up and finish up octree generation. And I got a great idea on how to collect thread statistics. But first and foremost, got to find a job. Still nothing.