cleaned up root directory
Travis became worthless as soon as we made GCC 4.8 mandatory. Travis' whole point was for me to avoid having to compile twice for GCC and Clang, which is now a moot point. Not that compiling is able to tell you much about whether it's going to run fine.

Still can't put much time. Got 2 weeks to find what to do and where to live. Current prospects include: an agonizing free 3 hour journey(one way) by train to a non-exotic location for an interview and NULL. Stumbled upon something yesterday, will sharpen my CV and Cover letter and send it tommorow. And some goddamned fuckers have yet to respond after 3 months. I'll probably see an email in December asking if I'm up for an interview.

Whatever, enough rambling, I'm close to 0.2, just need to fix up the octree generation. I think the problem's in the initial dimension.