graph: reenabled object selection, random code tweaks
I'll probably define a struct containing all camera details and send that over to graph_draw_scene. For now this'll do. Now I'm scratching my head over those octrees. For some reason as soon as the root octree leaves the scope of the object insertion function it's gone and gone for good. What the hell. Regradless of whether I malloc it(altough I shouldn't for such a small object).