physics-aux, in_molecule: fixed importing
Fixed now. I am not at all regretting allocating memory for each cell now. Not in the slightest. Trees can grow quite a lot, considering the additional layers required if the objects are far from the origin of the root octree. It'd probably be wise to move that origin to the centre of mass. Anyway, you can now play around with molecules and proteins in octrees. I am thinking of using a lua script to read 'em, much in the same way physics-aux is right now, but we'll see.