physics-all, graph: implemented per-thread statistics
Interesting, as the entropy falls down so do the total amount of octrees. As expected. I still need to figure out how to thread the octree construction. The thread time statistics are back as well. Using a multidimensional array was not wise, turns out it's faster to just print them sequentially.