physics-bh: set malloc's fastbin size to that of an octree
And a bunch of other tweaks. Going a few more chars over 80 is perfectly fine. Anyway, this should work better. Should. I'll run two instances and sees which wins. Survival of the fittest. Now if we could automate these tests and make efficient code write itself we'd be in the Matrix by 10 years. Kek.