physics, graph: object removal fixed, tweaks to scene drawing
Works fine, can be fun to remove the first 2000 objects to get a donut. Just make sure the physics' off, otherwise it'll get unhappy. I have an idea on how to do depositions on surfaces: allocate memory for the objects beforehand, put them in there and keep'em until a certain time, then reinitialize the physics. But first, gotta think of a nice demo for the BH algorithm. And thread it.