physics-bh, graph: prep work for multithreading bh creation
Most of it was done a few days ago. What did I do in between now and then? Thinking how to thread the beast. I've settled on a single octree per thread, now I need to decide how exactly to determine which thread to get what. "The Day of the Triffids" is a nice book though, definately a nice read.