make, physics-bh: updated default flags, updated distribution f-n
Automake will autoinsert "-g -O2" and the way to stop this is by setting them to nil. This also stops us from using environmental variables to influence them but whatever, maybe I'll add that as an argument if I really need it. I'm still figuring out exactly how threading should work and I believe I got myself a few ideas. About how threads should decide which object to take - just recursively insert the object until the depth at which the thread octree is at and then do a pointer comparison. I was thinking of using a function to determine whether an object lies inside the boundary(which is still not a bad idea, recursions vs calculations).