physics-bh: rewrote thread distribution function
Algorithm's finally finished, that's what I worked on for the last couple of days. The qsort thread compare function does not work and am working on a mock to see what the hell I need to dereference to get the correct number. Also, the memory is not freed so I'll have to write another decimate function to wipe it clean. Sure polymorphism would help alot since the octrees for the cells and the threads are basically identical, but I'm not ready to sell my soul. After the whole BH odyssey has been completed I'll do a demo(got some nice ideas), tag a 0.2 commit and spread the word again. Working with recursive functions has been frustrating but hey, at least I got(or like to think I've got) a ton of street cred now.