Switched entire build system to Waf
Autotools were a nightmare to work with. I've seen rigid, unflexible languages like COBOL but man, using autotools was like trying to bend a 4 inch thick steel pipe in a spiral. Absolutely unusable. Horrible. And it was so slow. For a single change to any makefile you have to run autoreconf and then the configure script. I bet I could write a better build system in 48 hours in COBOL. And I don't even know COBOL. Waf seems to be awesome, so yep, gonna use it instead. Still need to fix the hack at the very end of the script. Also removed all and any molecule files. Still in the git repo, so might have to purge it as well. Anyway, will hopefully finish the threading of the octree generator. Much easier to work with now. That monster of a system was an absolute pain to implement in the first place. Also, file2string was copied from mpv. Because the main philosophy of this project seems to be: if mpv does something, it's probably the right thing to do so copy it.