physics-bh: fixed multithreading
Yes, in all its glory, it's more or less done. Still surprised it only took 500 lines of code. Shouldn't be as it's by far the largest file in the entire tree. Or maybe I'm that pedantic about splitting code. Whatever, still need to figure out exactly why it crashes when I move the root's position EXACTLY rather than like now. But it's more or less done. And it's as beautiful as I imagined it'd be. Just look at how much faster it is than n-body, holy shit. Turn that ratio up a bit and watch it devour hundreds of thousands of objects in a matter of seconds. Oh yeah, and I need to see why some versions of gcc or clang vomit in disgust from my use of vectors.