physics-bh, graph: use linear scoring algorithm, increase seperation
It's not faster than what we had before but it's not slower. Also reduces memory. Basically, the cleanup function will reduce an octree's score by one and if an object comes in and populates it, it's going to have it's score increased by one. So it'll not drain it as long as objects are in, rather than as previously where if it had an object it would reset its score. Also, the lifetime's going to be increased by one on initialization, but that's hardly going to make a difference.

Some clipping occured with huge octree flux, so adjusted for that as well.