physics-bh, parser: option added to adjust single assignment, improvements
Found the bug. Or what's causing it. You see, the function to determine the halfdim goes a little something like this: root_halfdim/pow(2, depth). Plot that and you'll see it fall down to 0 real fast. And if the halfdim is 0 the octrees spawn on the same coordinate since their position has not been advanced. Currently thinking how to fix it. There should never be a need for more than 30 octrees as per the current settings in simconf, so I'll play around and see what happens.