physics-bh: finalized
You bastard. Took me a while to hunt this one. Turns out the precision was not enough on the halfdim variable. Since it decays ridiculously fast as a function of depth at some point it reached exactly 0 and then it's game over man. The octrees started spawning in the same position and the objects never get inside an octree. Fixed. Also renamed v4sd to a much more sensible and standard vec3. I need to get the screenshot function working, maybe come up with an event synchronization system for doing snapshots at the end of every timestep and make demos.