parser, in_file: optimize, implemented model scaling
Now it's much easier to add stuff to the importer. And I made the parser such that it only reads and executes the Lua script once. So if you have tons of objects it'll be a bit faster. I might add support for rotation soon. Need to get the screenshot function to work first.