parser: expose entire object array to Lua
Works well. The performance increase is not as bad as I thought it would be. I might create an overcomplicated API to deal with changing the object's attributes.

Though one should note the structure of objects in spawn_objects and obj in run_on_timestep are different. The latter has the position and velocity in terms of tables.