release: 0.3 rc
Lots of work went here. I was planning to wait until I got RK4 as an algoritm, but just as I suspected, it's a piece of shit. Having to construct 4 octrees is bullshit. Having to calculate something 4 times is beyond reasonable. Velocity Verlet with a small timestep's where it's all at.