all: renamed object datatype, removed perl build dependency
data was mistifyingly mysterious. What data? Pointer to data? Const char array? Flying monkeys? Make it clear it's a physics object. File2string was the only perl script/thing used, but since we require python to build might as well change it. Thanks to the gentoo guys for making it python2 compactible and the mplayer2 guys for actually writing the script. It seems mpv changed it word for word to perl for some reason.