rc4(?)
Working on a lot of things at the moment. Goddamned piece of shit. I can't get the fucking last object to update. Is Lua playing retarded? It apprently quits right after it sets the last object's position vector. Why? How is any of this different? If update all objects it works. Kinda. What the fuck.