physics-t_lua: Added multithreaded Lua-only algorithm
It's not pretty. We ask Lua to create a context for every thread and we give the exec f-n a range of objects with which it can operate. It's up to the thread to respect that. There is a problem where we go over the memory limit for Lua, still working on solving that. And there's still the huge problem of needing an extra object at the end else we can't parse all objects off the stack. And the performance gains are probably minimal. But it's multithreaded Lua, man. Bet any Lua developer would laugh at the way it's implemented though.