fixed: stepping forward, stepping back, MODE_POINT, physics-bh
Stepping back was incorrect, got memcpy a bit confused. Was looking at st.size in output.c which is in bytes. MODE_POINT didn't use white. Physics-bh had its threading broken for like a few weeks now. Forgot to test. Problem was we need to allow for one variable to be negative. Forgot which. Yeah, I am going to sacrifice my external HDD and put something on it, go to a more capable computer and do some screencapturing. I figure an i7 with 4 cores and 8 threads ought to be enough to do some stuff with around 100 000. Will probably use the png dump function to get a snapshot every step.