waf: disable default native optimizations, relax Clang version to 3.4
I suspect its causing a problem. Suspect being in present tense because I need to confirm it tommorow. On high-end machine with AVX and AVX2(heh, high-end compared to my B820, basically anything since 2010 or so) I think GCC optimizes vectors and creates incorrect machine code. Either that or Ubuntu's GCC is shit compared to Debian's. Possibility here is real, this is C extensions we're dealing with.

Clang on Ubuntu seems to piss itself and die at the parser. When we assign objects from the buffer to the address. Tried memcpy. Didn't work. I think it does a worse job at making code because it cannot even assign a vector. I'll see tommorow.

Clang 3.4 compiled this thing so I guess it should work. Need to test tommorow.