parser, physics: fixed buffer switching(under Clang), added coord check Lua f-n
Needed such a function to leave it to the user to decide what to do with conflicting objects. It lists them up nicely. Also Clang broke back buffering for some reason, so fixed it. Turns out we're now correct so a buffer position of 0 indicates (present) and a buffer position of step_back_buffer_size is the last.