physics, manpage: FIXED buffer segfaults, completed manual
Yes, FIXED the buffer. Completely this time. No idea why Clang crashes if we do it directly, but this is THE CORRECT AND ONLY way to do it. All the indecies are below/on the allocated space, so not sure why shuffling a bunch of 8 byte pointers would cause the whole thing to crash. Damn, pointers are huge nowadays, aren't they? I remember back in my day we had...