WIP/Broken: physics_barnes_hut viscosity
Not working yet. In fact build is broken, but I need to start digging through the stuff I did a month ago to resume and finish a few things. Revert to previous commit if you need something working.

As for what made me stop doing this for a month, come on IRC anytime you'd want to hear me rambling.