physengine: change name to \'glauss\'
Finally got around to thinking of a good name. We'll see how it goes. \'physengine\' has always been just a temp name because I didn't know what to call it when I started it. If anyone's wondering about the origin of the name, just think of it as some randomly generated gibberish. Or whatever, you're free to interpret it as glau + gauss. If the name bugs you, go to the new IRC channel or put an issue up to complain.