sighandle: switch to a dynamic quit handle system
No idea why I didn't think of it when I first write sighandle. W/e, if someone starts abusing the system I'll add extra code to remove and reshuffle the array. But it's a pain and I'll have to use locking because it can be called from other threads, too much work so for now I won't bother.