server: able to run simulations
Still far off from actually serving as a server, but for now it's a nice headless program. Except that the input_thread has to be disabled because I screwed up by putting the damn graph settings in there too, which aren't even included in the server and will never be. I need to get around to seperating that, which is really the last thing preventing building the program without SDL2, libpng, freetype or opengl.

On a sidenote, what's with all the people starring this project all the suddent? Surely it isn't the change of name, it used to have a horrible descriptive name and now the name isn't even descriptive anymore.